Move OneWireConfigEntry type definition
